# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InviteCreate": {
        "properties": {
          "expires_in_hours": {
            "anyOf": [
              {
                "maximum": 8760.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 168,
            "title": "Expires In Hours"
          },
          "max_uses": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          }
        },
        "title": "InviteCreate",
        "type": "object"
      },
      "InviteCreated": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "community_id": {
            "format": "uuid",
            "title": "Community Id",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          }
        },
        "required": [
          "code",
          "community_id",
          "expires_at",
          "max_uses"
        ],
        "title": "InviteCreated",
        "type": "object"
      },
      "InviteRedeem": {
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "InviteRedeem",
        "type": "object"
      },
      "InviteRedeemed": {
        "properties": {
          "community_id": {
            "format": "uuid",
            "title": "Community Id",
            "type": "string"
          },
          "community_name": {
            "title": "Community Name",
            "type": "string"
          }
        },
        "required": [
          "community_id",
          "community_name"
        ],
        "title": "InviteRedeemed",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Openvoice API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/communities/{community_id}/invites": {
      "post": {
        "operationId": "create_invite_api_v1_communities__community_id__invites_post",
        "parameters": [
          {
            "in": "path",
            "name": "community_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Community Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
    "/api/v1/invites/redeem": {
      "post": {
        "operationId": "redeem_invite_api_v1_invites_redeem_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRedeem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteRedeemed"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Redeem Invite",
        "tags": [
          "invites"
        ]
      }
    },
...
  }
}